demos/requirements.txt
python-dotenv
langchain
langchain-core
langchain-openai
langchain-aws
pydantic
langsmith
langgraph
requests

langchain-pinecone
